Completion
Problems fixed:

When I attack an enemy it can go into minus point numbers, but it does
not say that the enemy is defeated, and it lets me persist attacking it.
(Now FIXED)
When an enemy has been defeated we need the player’s points to increase.
The points should be based on the enemy’s health points. (Now FIXED)
I need to have a random option so that there is a half chance that the
enemy can attack the player first. (Now FIXED)
When I pick up a spell and use it the spell is removed from my bag, but
if I stay in the same room I can pick up the spell I’ve already used.
This shouldn’t happen. Once the spell is placed in the bag it should be
removed from the room. (Now FIXED)
In Female Bathroom one of the directions is ‘name’. Please change this
so that it shows the valid direction.